[ex77][fmha_gen] Support KV cache > 4GB (i.e., int32 -> int64)
Test:
```
./77_blackwell_fmha_gen_fp16 --b=256 --h=8 --h_k=1 --k=65536 --d=128 --iterations=2 --cache-only --verify
./77_blackwell_fmha_gen_fp8 --b=256 --h=8 --h_k=1 --k=65536 --d=128 --iterations=2 --cache-only --verify
```

Before:
```
Failed to initialize the CUTLASS kernel. Last CUDA error is: no error
terminate called after throwing an instance of 'std::runtime_error'
  what():  Failed to query occupancy.
  Aborted (core dumped)
```

After: run correctly.

+++ examples/77_blackwell_fmha/77_blackwell_fmha_gen.cu
@@ -338,7 +338,7 @@ struct ExampleRunner {
 
   using StrideQ = Stride<_0, _1, Stride<Stride<int, int>, int>>;
   using StrideNewK = Stride<_0, _1, Stride<Stride<_0, int>, int>>;
-  using StrideCacheK = Stride<int, _1, Stride<Stride<_0, int>, int>>;
+  using StrideCacheK = Stride<int, _1, Stride<Stride<_0, int64_t>, int64_t>>;
   using StrideNewV = StrideNewK;
   using StrideCacheV = StrideCacheK;
   using StrideO = StrideQ;
@@ -478,7 +478,7 @@ struct ExampleRunner {
 
     stride_q = make_stride(_0{}, _1{}, make_stride(make_stride(options.d, options.d * size<3,0,0>(result)), options.d * size<3,0>(result)));
     stride_new_k = make_stride(_0{}, _1{}, make_stride(make_stride(_0{}, options.d), options.d * size<3,0,1>(result)));
-    stride_cache_k = make_stride(options.d * size<3,0,1>(result), _1{}, make_stride(make_stride(_0{}, options.d), options.d * size<3,0,1>(result) * get<1>(result)));
+    stride_cache_k = make_stride(options.d * size<3,0,1>(result), _1{}, make_stride(make_stride(_0{}, static_cast<int64_t>(options.d)), static_cast<int64_t>(options.d) * size<3,0,1>(result) * get<1>(result)));
 
     stride_new_v = stride_new_k;
     stride_cache_v = stride_cache_k;
